+++ packages/sources/svelte/src/components/navigation/VtmnBreadcrumbItem/test/VtmnBreadcrumbItemWithSlot.svelte
@@ -1,0 +1,5 @@
+<script>
+  import VtmnBreadcrumbItem from '../VtmnBreadcrumbItem.svelte';
+</script>
+
+<VtmnBreadcrumbItem {...$$restProps}>Test breadcrumb item</VtmnBreadcrumbItem>